Removed unused classes, renamed files to match class names
