--- pstatus-notifications-workflow-ktor/src/main/kotlin/model/CheckUploadResponse.kt
@@ -1,10 +1,0 @@
-package gov.cdc.ocio.processingnotifications.model
-
-
-/**
- * The POJO class which is used for mapping results from the db
- * @property id String
-  */
-data class CheckUploadResponse(
-    var id: String = ""
-)
