+++ pstatus-notifications-workflow-ktor/src/main/kotlin/model/CronSchedule.kt
@@ -1,0 +1,19 @@
+package gov.cdc.ocio.processingnotifications.model
+
+import java.time.OffsetDateTime
+
+/**
+ * Raw cron schedule and its human-readable form.
+ *
+ * @property cron String?
+ * @property description String?
+ * @property lastRun OffsetDateTime?
+ * @property nextExecution String?
+ * @constructor
+ */
+data class CronSchedule(
+    val cron: String?,
+    val description: String?,
+    val lastRun: OffsetDateTime?,
+    val nextExecution: String?
+)
