--- pstatus-notifications-workflow-ktor/src/main/kotlin/model/ErrorDetail.kt
@@ -1,12 +1,0 @@
-package gov.cdc.ocio.processingnotifications.model
-
-/**
- * Error Detail class
- * @property description String
- * @property count Int
- */
-data class ErrorDetail(
-    val description: String,
-    val count: Int
-)
-
