--- pstatus-notifications-workflow-ktor/src/main/kotlin/model/Subscription.kt
@@ -1,9 +1,0 @@
-package gov.cdc.ocio.processingnotifications.model
-
-/**
- * Data stream errors notification unSubscription data class which is serialized back and forth  from graphQL to this
- * service.
- *
- * @param subscriptionId String
- */
-data class DataStreamTopErrorsNotificationUnSubscription(val subscriptionId: String)
